Rename files, manually test the file-checker
