+++ tests/file_checker_test.py
@@ -1,5 +1,5 @@
 
-from app.repo_reporter import *
+from app.file_checker import *
 
 def test_read_filepaths_from_file():
     filepaths = read_filepaths_from_file("tests/mocks/filepaths.csv")
